test(protocol-contracts): separate task tests from regular tests
