+++ protocol-contracts/staking/test-tasks/deploy/protocolStaking.test.ts
@@ -1,6 +1,6 @@
-import { getProtocolStakingProxyName } from '../../../tasks/deployment';
-import { getRequiredEnvVar } from '../../../tasks/utils/loadVariables';
-import { expectContractDeployed } from '../../helpers/addressCheck';
+import { getProtocolStakingProxyName } from '../../tasks/deployment';
+import { getRequiredEnvVar } from '../../tasks/utils/loadVariables';
+import { expectContractDeployed } from '../utils';
 import { expect } from 'chai';
 import { ethers } from 'hardhat';
 import hre from 'hardhat';
