+++ protocol-contracts/staking/test-tasks/setters/deposit.test.ts
@@ -1,4 +1,4 @@
-import { getRequiredEnvVar } from '../../../tasks/utils/loadVariables';
+import { getRequiredEnvVar } from '../../tasks/utils/loadVariables';
 import { getOperatorStakingContractsFixture } from '../utils';
 import { loadFixture } from '@nomicfoundation/hardhat-network-helpers';
 import { expect } from 'chai';
