+++ protocol-contracts/staking/tsconfig.json
@@ -20,5 +20,5 @@
   },
   "exclude": ["node_modules"],
   "files": ["./hardhat.config.ts"],
-  "include": ["src/**/*", "tasks/**/*", "test/**/*", "deploy/**/*", "types/"]
+  "include": ["src/**/*", "tasks/**/*", "test/**/*", "deploy/**/*", "types/", "test-tasks"]
 }
